fix preprocessor logic with HAVE_STAT
The _WIN32 check is inside a preprocessor branch that has already
checked that _WIN32 is not defined.

On the other hand, if HAVE_STAT is not available, then the function
does not have an alternative method to check that a file exists. The
not-apple-not-windows branch should have stat() available, so this is
a very unlikely situation.